Move common functions to utils file
